requirements.txt
django
djangorestframework
python-decouple
djangorestframework-simplejwt
pillow
twilio
phonenumbers
django-jazzmin
drf-material
icecream
uvicorn
gunicorn
whitenoise
django-debug-toolbar
django-cors-headers
django-extensions
django-import-export[all]
eskiz_sms
environs
pytest
pytest-factoryboy
drf-standardized-errors
psycopg
django-filter
django-ulid
django-redis
django-ratelimit
django-schema-viewer==0.5.3
celery
flower
adrf==0.1.9
drf-yasg==1.21.10